+++ autoload.php
@@ -10,5 +10,5 @@ require_once __DIR__ . '/src/LatnIastToDeva.php';
 require_once __DIR__ . '/src/LatnIastToGujr.php';
 require_once __DIR__ . '/src/LatnIastTranscription.php';
 require_once __DIR__ . '/src/BrahmicToLatnIast.php';
-require_once __DIR__ . '/src/LosslessTransliteration.php';
+require_once __DIR__ . '/src/Transliteration.php';
 require_once __DIR__ . '/src/DevaGujrConverter.php';